failure_assertions
