+++ crates/forge/tests/cli/main.rs
@@ -16,6 +16,7 @@ mod create;
 mod debug;
 mod doc;
 mod eip712;
+mod failure_assertions;
 mod geiger;
 mod inline_config;
 mod multi_script;
